+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `GetRealmReply` and `UpdateRealm`.
 - Add `GetDeviceRegistrationLimit[Reply]` to retrieve the maximum
   number of registered devices per realm.
+- Add `GetDatastreamMaximumStorageRetention[Reply]` to retrieve the maximum
+  retention value for data in a realm.
 
 ### Changed
 - Bump Elixir to 1.15.7.